backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
matplotlib==3.7.1
networkx==3.3
numpy==1.24.3
pandas==2.0.0
Pillow==9.5.0
Pillow==10.3.0
